menambah permission artikel

+++ app/Http/Controllers/ArticleController.php
@@ -71,6 +71,7 @@ class ArticleController extends Controller
     }
 
     public function edit(Article $article){
+        abort_if(auth()->user()?->isNot($article->user), 401);
         return view('articles.edit', [
             'article' => $article,
             'categories' => $this->categories,
@@ -79,6 +80,7 @@ class ArticleController extends Controller
     }
 
     public function update(Request $request, Article $article){
+        abort_if(auth()->user()?->isNot($article->user), 401);
         $request->validate([
             'title' => ['required'],
             'body' => ['required'],
@@ -100,6 +102,7 @@ class ArticleController extends Controller
     }
 
     public function destroy(Article $article){
+        abort_if(auth()->user()?->isNot($article->user), 401);
         if ($article->picture) {
             Storage::delete($article->picture);
         }
